reconfigure some things
