+++ server/src/lib.rs
@@ -6,6 +6,7 @@ pub mod ledger;
 pub mod models;
 pub mod queries;
 pub mod router;
+pub mod processor;
 
 #[macro_use]
 extern crate postgres_derive;
