+++ server/src/router/mod.rs
@@ -29,7 +29,7 @@ impl Build for Router {
 
         Router::new()
             .merge(spa)
-            .route("/api/v1/:keyword", get(api::handler))
+            .route("/api/v1/:keyword", get(api::keyword::handler))
             .fallback(react_router_fallback)
     }
 }
